requirements/requirements.txt
# bcolz>=1.2.1
scipy>=1.6.2
numpy>=1.19.2
torchvision>=0.9.0
torch>=1.8.0
pandas>=1.2.4
matplotlib>=3.3.4
tqdm>=4.51.0
Pillow>=8.3.0
scikit_learn>=0.24.2